fix rate
