+++ src/store/UploadTransStore/UploadTransStore.ts
@@ -91,15 +91,4 @@ export default class TokenInfoStore {
     } {
         return this._repos;
     }
-
-    destroy(): void {
-        this._metaChangedReaction();
-    }
-
-    _metaChangedReaction: IReactionDisposer = reaction(
-        () => this.meta,
-        (...args) => {
-            log('Reaction', args);
-        },
-    );
 }
