+++ src/utils/formatMoney.ts
@@ -3,16 +3,18 @@ export const formatMoney = (number: string | number | undefined, decimal: number
     const idxDel = strNumber.indexOf('.');
     const mod = strNumber.slice(idxDel);
     const count = mod.search(/[1-9]/);
+    let isRate = false;
+
 
     if (count > 1 && Number(decimal) === 2) {
+        isRate = true;
         decimal = count;
     }
 
     const typeNumber = Number(number);
 
     const rounded = Number(typeNumber.toFixed(decimal));
-
-    return new Intl.NumberFormat('en-US', { style: 'decimal' }).format(rounded);
+    return isRate ? rounded : new Intl.NumberFormat('en-US', { style: 'decimal' }).format(rounded);
 };
 
 export const formatDiff = (diff: string | number): string => {
