+++ src/utils/useLocal.ts
@@ -11,14 +11,13 @@ export const useLocal = <T>(creator: () => T): T => {
 };
 
 export interface ILocalStore {
-    destroy(): void;
 }
 
 export const useLocalStore = <T extends ILocalStore>(creator: () => T): T => {
     const store = useLocal(creator);
 
     React.useEffect(() => {
-        return () => store.destroy();
+        return () => {};
     }, []);
 
     return store;
